add test_interface.py
